Massive zone file (de)serialization rewrite
Wow, okay, commit message time. Um.

The tl;dr is that we're switching to MessagePack for serializing and
deserializing zones. This is primarily due to the fact that DNS-encoded
messages have a maximum length, we don't want to cap out zone files, and
I don't want to understand how AXFR messages are continued.

All name labels in the zone are stored at the beginning, and names
within the zone are converted to vecs of indexes so that the zone file
size is smaller and *in theory* we use less memory to store the zone
because of deduplication logic. In practice this doesn't happen yet
because we're naïvely converting names from our own implementation to
TRust-DNS's and so every label in every name allocates.

We've reimplemented Record and RData types to make it simpler to build
APIs around these types, and then implemented the one-way conversions
into TRust-DNS's Record and RData types.

There's still a lot of room for improvement here, but right now it's
more important to start building the features that we care about for
pepbut now, and work on optimization later.