Switch from rmpv to rmp
Instead of building up an rmp::Value and then encoding it, or decoding
into an rmp::Value and destructuring it, implement our own readers and
writers using the direct rmp decoders and encoders.

In order to do this well, I've modified the zone format to move the
labels at the end. We still need the labels first when reading, so we
also write the byte offset from the end of the zone file containing the
location where the label array starts.