Temporarily remove pepbutd and clean up internals
The parent of this commit is the closest I've come to building a
compromise between how I want my API to look and what will work for
TRust-DNS from the perspective of its server.

I've decided that TRust-DNS is not designed for mucking about in its
internals, and that my opinions on how this particular DNS server should
be built is different from the opinions made in TRust-DNS.

I'm leaving a reference to the project in README.md as a lot of the data
types we have so far are heavily inspired by it. I suspect a lot of the
code to parse requests and write responses on the wire is going to be
inspired from it as well.

Additionally, because apparently I don't git commit enough, I've added
clippy's pedantic warnings and have gone through and cleaned them all
up, including writing documentation for everything.