Store records as HashMaps; require nightly rustc
When we eventually implement record lookup, storing the records as a
HashMap indexed on the name and the record type will be the only
efficient system.

Add convenience methods to Zone that make modifying the set of records
ergonomic. To implement Zone::iter, we are relying on the unstable
conservative_impl_trait feature. To more easily implement Zone::remove,
we are relying on the unstable vec_remove_item feature.

Since we are taking a dependency on the nightly compiler, remove the
nightly feature for benchmarks.