Remove separate Label type
This makes the implementation of Names somewhat less confusing.

Name::from_str now skips IDNA conversion if all the bytes in the name
are alphanumeric or a hyphen (excepting the first byte being an
underscore for SRV records). This requires adding an EmptyLabel variant
to ParseNameError since IDNA conversion is not causing an error here
anymore.