Use bytes::{Buf, BufMut} instead of byteorder
bytes has methods similar to byteorder's that only work on buffers, thus
removing the need to handle std::io::Errors that are chained all the way
from Read/Write.

Because bytes does not grow the underlying Cursor<Vec<u8>> buffer when
writing to the end, remove the Cursor, since we only need to know the
position of the cursor when it's already at the end of the buffer.