Remove FIXME for mapping to () in nsd.rs
We already map all the values to () in the futures themselves (logging
errors to the console all the while).

Items and Errors in the Select<A, B> future are tuples returning the
value from the future that finished first, and then the future that
hasn't finished yet. If either future stops, we want to take down the
whole program with it so that the init system can restart us.